Fix reference to undefined type name.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/trunk/crossfire@4576 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ socket/request.c
@@ -1022,7 +1022,7 @@ static uint16 add_head(object *ob, int sx, int sy, int p_layer)
 {
     object *head;
     int bx, by, i;
-    uint face_num=0;
+    uint16 face_num=0;
 
     if (ob->head) head = ob->head;
     else head = ob;
